Ensure dependency between inner loops.
For large strides, the inner loop can be very short (2-4 accesses). If
there is no dependency between loops, the first access of the next loop
can happen in parallel with the last access of the previous loop, making
the total latency shorter than the sum of all accesses.

Ensure a dependency on ``nextstep'' between loops. Also change the
cancelling loop to have the same amount of work.